requirements.txt
fastapi
sqlalchemy
psycopg2-binary
fastapi_mail
python-jose